# XpertGroup.Servicio/Controllers/ServicioApi.cs
/*
 * XpertGroup Prueba
 *
 * Este es el api que se podra consumir desde la aplicacion 
 *
 * OpenAPI spec version: 1.0.0
 * Contact: jgonzoe@gmail.com
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */

using IO.Swagger.Attributes;
using IO.Swagger.Models;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using System;

namespace IO.Swagger.Controllers
{
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class ServicioApiController : ControllerBase
    { 
        /// <summary>
        /// envie los requerimientos del servicio expuesta
        /// </summary>
        
        /// <param name="body">Todos los parametros requeridos para enviar el servicio</param>
        /// <response code="405">Invalid input</response>
        [HttpPost]
        [Route("/xpertGroup/solicitud")]
        [ValidateModelState]
        [SwaggerOperation("CrearSolicitud")]
        public virtual IActionResult CrearSolicitud([FromBody]Solicitud body)
        { 
            //TODO: Uncomment the next line to return response 405 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(405);


            throw new NotImplementedException();
        }
    }
}